python_backend/pyproject.toml
[project]
name = "mini-aiman-agent"
version = "0.1.0"
description = "Minimal LiveKit agent powering the Mini AiMan holographic demo."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0",
    "livekit-agents[openai,silero,blingfire]>=1.2.8,<2.0",
    "openai>=1.58",
]

[project.scripts]
mini-aiman-agent = "main:main"

[tool.uv]
cache-keys = ["openai"]
